+++ android/app/src/main/java/com/budgetblocks/MainActivity.java
@@ -1,4 +1,4 @@
-package com.basicbudgets;
+package com.budgetblocks;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactPackage;
@@ -15,7 +15,7 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "BasicBudgets";
+        return "BudgetBlocks";
     }
 
     /**
@@ -27,14 +27,14 @@ public class MainActivity extends ReactActivity {
         return BuildConfig.DEBUG;
     }
 
-   /**
-   * A list of packages used by the app. If the app uses additional views
-   * or modules besides the default ones, add more packages here.
-   */
+    /**
+     * A list of packages used by the app. If the app uses additional views
+     * or modules besides the default ones, add more packages here.
+     */
     @Override
     protected List<ReactPackage> getPackages() {
-      return Arrays.<ReactPackage>asList(
-        new MainReactPackage()
-      );
+        return Arrays.<ReactPackage>asList(
+            new MainReactPackage()
+        );
     }
 }
