+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'BasicBudgets'
+rootProject.name = 'BudgetBlocks'
 
 include ':app'
