+++ ios/BudgetBlocks.xcodeproj/xcshareddata/xcschemes/BudgetBlocks.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "BasicBudgets.app"
-               BlueprintName = "BasicBudgets"
-               ReferencedContainer = "container:BasicBudgets.xcodeproj">
+               BuildableName = "BudgetBlocks.app"
+               BlueprintName = "BudgetBlocks"
+               ReferencedContainer = "container:BudgetBlocks.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "BasicBudgetsTests.xctest"
-               BlueprintName = "BasicBudgetsTests"
-               ReferencedContainer = "container:BasicBudgets.xcodeproj">
+               BuildableName = "BudgetBlocksTests.xctest"
+               BlueprintName = "BudgetBlocksTests"
+               ReferencedContainer = "container:BudgetBlocks.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -47,9 +47,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "BasicBudgetsTests.xctest"
-               BlueprintName = "BasicBudgetsTests"
-               ReferencedContainer = "container:BasicBudgets.xcodeproj">
+               BuildableName = "BudgetBlocksTests.xctest"
+               BlueprintName = "BudgetBlocksTests"
+               ReferencedContainer = "container:BudgetBlocks.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -57,9 +57,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BasicBudgets.app"
-            BlueprintName = "BasicBudgets"
-            ReferencedContainer = "container:BasicBudgets.xcodeproj">
+            BuildableName = "BudgetBlocks.app"
+            BlueprintName = "BudgetBlocks"
+            ReferencedContainer = "container:BudgetBlocks.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
    </TestAction>
@@ -77,9 +77,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BasicBudgets.app"
-            BlueprintName = "BasicBudgets"
-            ReferencedContainer = "container:BasicBudgets.xcodeproj">
+            BuildableName = "BudgetBlocks.app"
+            BlueprintName = "BudgetBlocks"
+            ReferencedContainer = "container:BudgetBlocks.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -96,9 +96,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BasicBudgets.app"
-            BlueprintName = "BasicBudgets"
-            ReferencedContainer = "container:BasicBudgets.xcodeproj">
+            BuildableName = "BudgetBlocks.app"
+            BlueprintName = "BudgetBlocks"
+            ReferencedContainer = "container:BudgetBlocks.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
