+++ ios/BudgetBlocks/AppDelegate.m
@@ -36,13 +36,15 @@
   /**
    * OPTION 2
    * Load from pre-bundled file on disk. The static bundle is automatically
-   * generated by "Bundle React Native code and images" build step.
+   * generated by the "Bundle React Native code and images" build step when
+   * running the project on an actual device or running the project on the
+   * simulator in the "Release" build configuration.
    */
 
 //   jsCodeLocation = [[NSBundle mainBundle] URLForResource:@"main" withExtension:@"jsbundle"];
 
   RCTRootView *rootView = [[RCTRootView alloc] initWithBundleURL:jsCodeLocation
-                                                      moduleName:@"BasicBudgets"
+                                                      moduleName:@"BudgetBlocks"
                                                initialProperties:nil
                                                    launchOptions:launchOptions];
